Moved API server functions to separate files
