+++ client/package.json
@@ -6,15 +6,16 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^18.2.0",
-    "react-cookie": "^4.1.1",
-    "react-dom": "^18.2.0",
-    "react-scripts": "^5.0.1",
-    "web-vitals": "^2.1.4",
     "bootstrap": "^5.1.3",
     "cookie-parser": "^1.4.6",
+    "net": "^1.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
-    "react-router-dom": "^6.4.0"
+    "react-cookie": "^4.1.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.0",
+    "react-scripts": "^5.0.1",
+    "util": "^0.12.4"
   },
   "scripts": {
     "start": "react-scripts start",
